requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.2
Automat==0.7.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.2
certifi==2019.3.9
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-cors-headers==2.5.1
django-ipware==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.3.0
msgpack==0.6.1
psycopg2==2.7.7
PyHamcrest==1.9.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
soupsieve==1.8
static3==0.7.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
zope.interface==4.6.0
